celery + rabbitmq integration seems to be working now.  In order to test it out we need to make sure three services are running: 1. celerybeat (python manage.py celerybeat) 2. celeryd (python manage.py celeryd) 3. rabbitmq (/etc/init.d/rabbitmq-server if installed properly as a system level package)
one issue with the current implementation is that /var/log/rabbitmq/rabbit.log gets
huge with the every-second tick.  It looks like rabbitmq logging is not
configurable to the degree where we can specify a loglevel and have it not log INFO
events.  I don't particularly want to disable the rabbit.log but we may have to do
so to deal with the enormous number of 1s pings we're generating from a 1s
celerybeat schedule.