all api post endpoints should now expect to receive form posts from Sencha Touch and respond with a success flag and a custom payload in JSON.
fixed a number of bugs with the Comment model and removing some unnecessary cruft
from the core models (DataValue in particular is unnecessary - its foreign key to
Experiment was redundant and causes possibilities for update / data integrity
issues).  This might be the case for round_data and group since those are also tied,
should consider refactoring that more in the future.