renaming INSTRUCTIONS templates to GENERAL_INSTRUCTIONS and reserving INSTRUCTIONS template for the actual round instructions, set by an instructions round.
Current flow is to have initial WELCOME round kick off the GENERAL_INSTRUCTIONS, which leads to PRACTICE_ROUNDS.

After the practice rounds are over a new instructions round starts which should have an appropriate template_name to
refer to the correct treatment instructions (e.g., instructions-treatment-a.html or instructions-treatment-b.html,
etc.).  The treatment instructions template should be rendered by the server side and added to the view model.