changing control flow:
Previously, the last participant would receive a response with "all participants are ready" at which point it would
notify the realtime server which notifies the experimenter which notifies the django server to move to the next round

Now, participants just signal that they are ready to the django server and realtime server.  Every time the experimenter
receives a realtime participant_ready event it polls the check-ready-participants endpoint to see if all the
participants are ready and if so, advances to the next round.

The ideal situation would still be for the django server to advance the round directly and then tell the sockjs server
to update all connected clients.

TODO: wire up boundaries submit decision endpoint to set up participant_ready properly