fix ParticipantSignup.objects.with_attendance bug
Fixing copy-paste error and redirecting back to experiment session
signup when a session is full as there may be more sessions to select.
updating the subject pool test.

TODO: need to wire up waitlist into the front end, and allow for
waitlist confirmation as there may be other selectable sessions.