add ExperimentConfiguration.payment_information
Adding a dedicated field to ExperimentConfiguration to avoid updating
email templates whenever we run a new treatment and the payment
instructions change. Also refactoring neighborhood -> community to
better reflect the upcoming treatment options.

Also adding settings.ENVIRONMENT variable to distinguish between
production, staging, and dev environments. On the staging server we
still want to be able to send the nightly emails out for testing
purposes (but not to the subject pool participants).